+++ crates/config/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "config"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+authors.workspace = true
+
+[dependencies]
+config = { version = "0.15.3", features = ["toml"] }
+serde = { version = "1.0.219", features = ["derive"] }
+toml = "0.9.2"
+directories = "6.0.0"
